Move user tables to sensitive schema
